services/gmail_reader/requirements.txt
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
dotenv==0.9.9
google-api-core==2.25.0
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
httplib2==0.22.0
idna==3.10
oauthlib==3.2.2
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
python-dotenv==1.1.0
redis==6.2.0
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
soupsieve==2.7
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.4.0
